Refactored Error handling code.
I've been willing to change this for a while, but it took me longer than I though. This is a non trivial change, and it might break old code. The old error handling was very confunsing.Major changes:

1) Users can listen to BulkLoader.Error event ids for each item individually, or for the entire BulkLoader instance.
2) There's only on event type, and that is BulkLoader.Error, which will dispatch an  ErrorEvent ( before you had to listen to SecurityErrorEvent and IOErrorEvent separatly)
3) No more BulkLoaderErrorEvents
4) LoadinItem instances will keep a errorEvent member that will hold more information about the event.From error handlers, you can test which error it's with:
evt.target.errorEvent is SecurityErrorEvent or evt.target.errorEvent is IOErrorEvent
(before there was no way to retrieve error information)
5) Each error will fire as soon as it happens (before this change we waited untill all was loaded)
6) The BulkLoader.log will include a more useful error message (before this change all it said was there was an error)

All in all, this seems like a much better way to mange errors, but this might break older code.